добавили структуру dashboard
