+++ grape/res.qrc
@@ -4,5 +4,6 @@
         <file>img/password.svg</file>
         <file>img/login-background.png</file>
         <file>img/settings.svg</file>
+        <file>fonts/Comfortaa-VariableFont_wght.ttf</file>
     </qresource>
 </RCC>
